refactor : 코드를 세분화 시켜서 가독성을 높힙니다.
